test/wpt/jsdom/issues/3686-crash.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>3686 crash test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/jsdom/issues/3686 -->

<script>
"use strict";
test(() => {
  assert_equals(document.querySelector(':is([a],b):not(.c)'), null);
});
test(() => {
  assert_equals(document.querySelector(':is(a,[b]):not(.c)'), null);
});
test(() => {
  assert_equals(document.querySelector(':is(a,b):not(.c)'), null);
});
test(() => {
  assert_equals(document.querySelector(':is(a,b):not([c])'), null);
});
test(() => {
  assert_equals(document.querySelector(':is(a):not([b],.c)'), null);
});
test(() => {
  assert_equals(document.querySelector(':is(a,[b])'), null);
});
test(() => {
  assert_equals(document.querySelector(':is(a,[b]):first-child'), null);
});
test(() => {
  assert_equals(document.querySelector(':is([b]):not(.c)'), null);
});
</script>
